Move fetchArticle from component to the page
